requirements.txt
pygame
pyyaml
neat-python